Add files via upload
The Face Recognition and Age/Gender Detection Project" is a project that utilizes deep learning techniques and the OpenCV library to detect a person's face in live video from a webcam and determine their age, gender, and birth year. The project relies on well-trained pre-existing models for face detection and age/gender classification using deep learning techniques.

Furthermore, in addition to identifying faces, age, and gender, the project can also determine the birth year and display additional information based on the detected age. For example, the project can display historical information or cultural facts related to the detected age group, such as the year the person was born and the major event that occurred in that year.

This project can be used in various scenarios such as access control systems, statistical analysis of target audiences at events, analysis of marketing data in stores, and many other uses in diverse fields, including its use to celebrate national events such as the National Day by identifying the political era in which the person was born.